Abstract out LanguageSelector and LanguageOption
