+++ src/containers/LanguageOption.tsx
@@ -1,7 +1,7 @@
 import {ApplicationState} from '../types';
 import * as actions from '../actions/';
 import {connect} from 'react-redux';
-import LanguageSelector from "../components/menu/LanguageSelector";
+import LanguageOption from "../components/menu/LanguageOption";
 import {Dispatch} from "redux";
 
 
@@ -10,7 +10,9 @@ export function mapStateToProps(state: ApplicationState) {
 }
 
 export function mapDispatchToProps(dispatch: Dispatch<actions.Action>) {
-    return {};
+    return {
+        onSetLanguage: (language: string) => dispatch(actions.setLanguage(language))
+    };
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(LanguageSelector);
+export default connect(mapStateToProps, mapDispatchToProps)(LanguageOption);
